Splash Screen and Loading Screen Prefab
The Splash Screen is to be used as a way to initialize Singleton instances
like the LanguageManager.

Also added a small prefab for denoting that the game is loading a new
screen. Simply place it in the scene and it will put a rotating ying-yang
orb and a loading sign at the bottom right hand corner of the screen.
Simply destroy the object when the game is finished loading.