GameObject and Component Extension Functions
Added a large number of utility functions to improve retrieveing
Components.

GetIComponents and its variants are generic versions of
GetComponent that also can work with interface types.

SafeGetComponent is the same as normal GetComponent, but it logs an error
if a component that is expected is not found.

GetOrAddComponent will add a Component of the specified type if it does
not already exist.

Both extensions are created for GameObjects and Components alike.