Created Static Managers Injector
To mock as if the universality of the Static Managers object. Scene
post-process callback was added to inject one into the scene if it doesn't
already exist. It is a bit hacky in the sense that it has to look
explicitly for a object tagged with Static Managers tag (which was added
in this commit), but it works well.