CharacterData
Character Data is a general "hub" container for all relevant data to a
Character.

It contains no direct references so not all of the character's data is
loaded in as the CharacterData instance is loaded in itself.

Instead it contains a variety of useful properties for loading and
unloading these assets on command.

With that in mind, a generic Resource class was created to wrap strings so
that we weren't dealing with strings all the time, make the code more
declarative and readable. The Resource object is lazily loaded, only
loading the Resource when actually needed or explicitly told to do so.