Animation Parameter Wrappers
To avoid using strings, and ensure usage of hashes. Animation Parameter
information can be wrapped in Animation Parameter object instances that
act virtually as a normal C# property would. This abstracts away the need
to use strings and hashes as identifiers.

They are all directly editable from the editor, and appear as just a
normal string fields.