Character Restructure
Changed the character to act at the top level prefab. This changed it so
that the Animator component is expected to be a child of the Character
object.

Fixed the Player Indicator issue of spawning too low.

The console was having a lot of "Animator not initialized" warnings. If
this starts to become an actual issue, it will be added to the issue
tracker.