Networking Components
Found out that the new Unity 5.1 Networking components work just fine in
local single/multiplayer, and dont' throw errors if not in a networked
game. So now the Marisa Charaacter now has a NetworkIdentity,
NetworkTransform, and NetworkAnimator.

The Character class is now a network behavior and derives from
GensoNetworkBehaviour, an NetworkBehaviour derived alternative to
GensoBehaviour.