Refined and Fixed CameraController
Instead of using position based zoom, a field of view based on is used. To
prevent it from zooming in too much or too little or fisheye-ing, a
limitation is placed on the upper and lower bounds of the attainable FoV
values.

This limiation is made easier by the addition of a MinMaxRange attribute,
which has been added as well, and a custom PropertyDrawer alongside it.