Added lots of simple Materials to scene
Great aesthetic improvement over the bland grey looks from before.

Also fixed a bug with Marisa's animator not being updated when touching
the ground.

Added GroundCheck. A character component that takes care of whether the
character is grounded or not. This moves all logic involving groundedness
outside of the Character script itself.