Character Component Changes and Interfaces
CharacterComponents are no longer automatically attached via virtual
methods. This introduced too many extra empty stubs in each event. Now
they have to manually subscribe and unsubscribe to the events that are
relevant. Subscribe in Awake and Unsubscribe in OnDestroy.

Also added IDamageable and IKnockbackable interfaces for various objects
that can be knocked back or damaged.