Refactored Match-Stage Relationship
The previous implementation of Match and Stage was improperly created. The
Stage should simply be a data container behaiour and not need to deal with
the explicit spawning of the characters. Thus all logic for spawning and
starting a Match has been transfered from Stage.StartMatch to
Match.StartMatch.

To avoid extra data from being loaded when creating the Match, the
CharacterData references are used instead of Character references. The
Character prefab for actual gameplay will be loaded in when the Match
starts.

Likewise for better encapsulation of behavior, all basic Player Spawning
logic has been moved to Game.SpawnPlayer, which handles creating a prefab
from the CharacterData instance and creating the approriate player
indicator.

Match also no longer has InstantiateCharacter, relegating that logic to
Game.SpawnPlayer. Instead a OnSpawn method was created to allow the Match
to edit the Character instance created by Game.SpawnPlayer.

TestSpawnPlayer has been renamed to StockMatch in preparation for the next
step of creating a match for the standard Smash Stock matches.