Post Build File Copying
To add to the build automation, calling builds from the Genso menu will
also automatically copy all extra-build related files into the build
directory.

The folder is Assets/Post Build, which has three subfolders, each for one
platform to deploy to. Just put the files in there and they will be copied
to the approriate folder when built.