Including the UI project
It includes a scene with a basic Super Smash menu. When you are in the
character selection, you must use the arrows to control the cursor and
press space to click on the things. It is a test project so it is very
unpolished and it should include some bugs.