Changed the way Static Managers works
Static Managers is now changed so that the GameObject is set to static.
The object should never be destroyed, even by scene loads. In the final
build, the Static Managers object is bootstrapped in the splash screen
scene, and left there for the rest of the execution of the game. For
distinction, static manager objects created in this manner have an extra
(Bootstrapped) identifier on its name.

For the ease of development, alternative copies of static manager can be
placed development scenes. These ones are tagged "EditorOnly" and will be
removed on build. Should one stay in due to improper tagging, it will be
destroyed.

The max FoV of stage camreas has been increased to 75 because people were
saying it wasn't working in the demo on some points.