Seperated out logic for Grounding
Instead of the Character determining what is ground and what is not. The
stage instead tells the Character that a certain object is ground via the
Ground script.

Upon entry into a trigger or collision with objects with Ground scripts,
Characters are considered grounded, and recieve all effects related to
grounding like resetting the jumps. Upon exit, Characters are no longer
considered grounded, and will be treated as such.

The difference in implementation is important, as this allows any
arbitrary collider to be considered ground, versus the few tagged certain
tags.