Randomized Weighted BGM
Stages can now pull from a randomized pool of background music tracks,
weighted to the players' preferences, just like in Smash.

BGMGroup is a aggregation of Resource'd Background music clips, that also
stores the prefered weights in PlayerPrefs.

WeightedRNG is a generalized weighted selection object, it can be used for
a variety of uses, so it's class is decoupled from BGMGroup.

Stage Scripts can now take a reference to a BGMGroup, which it will use to
randomly select and play a track once loaded.