Lock Movement During Attack
Players will now not be able to move or jump while they are attacking.
Added a LockMovement method and a _canMove state variable to help with
this.
The method is somewhat sloppily called, and it doesn't play well with
aerial characters, but otherwise it works as intended.
*Also unchecked "Has Exit Time" from all animation transitions coming
from idle, since this was causing an (assumedly) unintentional delay
where other animations wouldn't play until the idle animation was
finished.