Preliminary Code for Fall Through Platforms
It triggers any time you are crouching on the platform and not only when
you tap it. That needs to be fixed.

Also fixed an issue with entering from the side of the platform.

Added Editor scripts for automatically generating a Box Collider trigger
for both Fall Through and normal platforms