Character Editor Window
Created a Custom Editor Window for editing Characters.

Added CharacterEditorData as a middleman. It contains all of the Editor
related information related to a character. It then uses this information
to generate runtime data, that is automatically generated inside a
Resources folder for access during gameplay.

Added EditorUtil and AssetUtil as utility classes for managing Editor
elements and Asset creation/use.