Required Character Components
This creates a way to require a Character to spawn with certain components
attached. On edit in the editor, all CharacterComponent types marked with
the RequriedCharacterComponentAttribute attribute will automatically be
attached.

If the components are deleted before runtime, they will be added upon
instantiation (in Awake), with their default values.

This could make for interesting setups, as the type search uses reflection
when the Character spawns to add all types. If types are included via a
network sent assembly, this could introduce some interesting additions.