Project Returned to "Working State"
The InitializeOnLoad call was glitching and deleting the CharacterData
since it cannot see it via Resources.Load until after the whole assembly
is loaded. Added a extra check to make sure the game wasn't transitioning
into Play Mode before generating a new CharacterData set.

The Prefab for Marisa was broken in that it couldn't move due to a
misplaced movement collider.

And finally the new AnimVarAttribute and Animator Extension methods seem
to be working well so far. Though some further testing is needed.