Movement, Damage, and Attacking Refactor. Statuses
The various funcitons related to Movement, Damage, and Attacking have been
moved out of the Character script and into their respective
CharacterComponents.

CharacterDamage in particular has been changed signifigantly. It now
allows for dynamic editing of the damage formula instead of being a direct
application of damage. It also has methods and fields related to healing
as well.

Two new types of CharacterComponent are added. Statuses and
CharacterActions.

CharacterActions represent an general action a character might do, like
Jumping or Attacking. CharacterActions have a built in restriction method
taht prevents the methods from being Invoked. Each has a Restrictions
event that allows adding Func<bools> for evaluation. For a invocation to
succeed, all of the delegates must return true.

Statuses are temporary (or permanent) alterations to the general state of
the Character. For example, Invincibiilty has been recreated as a Status.
It introduces a damage modifier that has highest priority (thus it always
runs last), and always returns a damage value of 0. Thereby removing all
damage from the Character. Each Status has a built in timer, triggered on
a call to StartStatus(). By default, statuses are permanent unless a call
to StartStatus() is made.