Merged Character Match Data and PlayerOptions
Single class named Player. These objects are created by the Match class
and are accessible via GetPlayerData. These contain data relevant to the
player's current status, such as if they are a CPU or not, what character
and pallete was chosen, or, if during the match itself, a reference to the
Character the player is using.