Resources Editor/Runtime Improvements
To reduce memory used when loading CharacterData, all complex or Unity
based data structures will be loaded in via Resources on demand.

Fixed a small bug causing Resource not to load Sprites properly, also
simplified the code a bit too.

Simplified the ResourcePathEditor too. Now looks like a regular old
object field. Only difference is that there is a indicator appended to the
label as to whether the object is in a Resources folder or not.