Added Editor Preview of Portraits w/Cropped
Added a preview to see the full and cropped portaits of included
Characters. Also added a notice at the bottom of the editor to tell if
the Character Data is placed in the proper location not. If the Character
Data is not placed in a Resources folder, an error will show saying that
it is not in the correct location.