Refactor/Consolidate UI Components
There were a lot of UI components that were accessing Character data
through the Player structure that also had a counterpart that only
interacted with the Character data directly. These were abstracted and
consolidated to match the necessary components and have it work with both
Player data and Character data.

Also added PlayerText to for the CSS.